Refactor tools into individual files
