+++ src/config.ts
@@ -1,0 +1,2 @@
+// List of glif IDs that should be available as tools
+export const GLIF_IDS: string[] = [];
